Fix several warnings.

+++ examples/SimpleDisplay/main.cpp
@@ -70,7 +70,7 @@ int main( int argc, char* argv[] )
   pangolin::Var<double> a_double_log("ui.Log_scale var",3,1,1E4, true);
   pangolin::Var<bool> a_checkbox("ui.A_Checkbox",false,true);
   pangolin::Var<int> an_int_no_input("ui.An_Int_No_Input",2);
-  pangolin::Var<CustomType> any_type("ui.Some_Type", CustomType(0,1.2,"Hello") );
+  pangolin::Var<CustomType> any_type("ui.Some_Type", CustomType(0,1.2f,"Hello") );
 
   pangolin::Var<bool> save_window("ui.Save_Window",false,false);
   pangolin::Var<bool> save_cube("ui.Save_Cube",false,false);
@@ -101,10 +101,10 @@ int main( int argc, char* argv[] )
     // Overloading of Var<T> operators allows us to treat them like
     // their wrapped types, eg:
     if( a_checkbox )
-      an_int = a_double;
+      an_int = (int)a_double;
 
     if( !any_type->z.compare("robot"))
-        any_type = CustomType(1,2.3,"Boogie");
+        any_type = CustomType(1,2.3f,"Boogie");
 
     an_int_no_input = an_int;
