+++ examples/SimplePlot/main.cpp
@@ -17,10 +17,10 @@ int main( int /*argc*/, char* argv[] )
   labels.push_back(std::string("sin(t)+cos(t)"));
   log.SetLabels(labels);
 
-  const double tinc = 0.01;
+  const float tinc = 0.01f;
 
   // OpenGL 'view' of data. We might have many views of the same data.
-  pangolin::Plotter plotter(&log,0,4*M_PI/tinc,-2,2,M_PI/(4*tinc),0.5);
+  pangolin::Plotter plotter(&log,0.0f,4.0f*(float)M_PI/tinc,-2.0f,2.0f,(float)M_PI/(4.0f*tinc),0.5f);
   plotter.SetBounds(0.0, 1.0, 0.0, 1.0);
   plotter.Track("$i");
